UT failure

+++ libs/features/mas/web-components/test/merch-card-product.test.html.js
@@ -56,7 +56,7 @@ runTests(async () => {
             const el =
               product.shadowRoot.querySelector(selector);
             if (!el) return 0;
-            return parseFloat(window.getComputedStyle(el).height);
+            return parseInt(window.getComputedStyle(el).height);
           })
           .join(',');
         return heights;
